Avoids infinite loop in normalization (uses readlink)
In normalize_prefix, the [[  ~= ]] test was always matching, resulting in an infinite loop.
I replaced the whole part with a simple call to readlink, which does the job
(with -m to be functionnaly the same, but -e would also be an option)

I'm not sure if readlink is available on every distrib/platform, so I left the old code to do
the work if readlink is not available.